Introduce frontmatters in step 2
Introducing liquid without introducing frontmatters can be a bit confusing and users of the tutorial run the risk of getting hung up on step 2, not realizing that their problem is addressed in step 3. A decent number of people are ending up here: https://stackoverflow.com/questions/29662793/using-liquid-tags-in-a-jekyll-page-not-working#comment94717164_29666241
Thanks for this great tutorial, though!